% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_accuracy.R
\name{adjusted_accuracy}
\alias{adjusted_accuracy}
\title{Compute adjusted accuracy given true and predicted values.}
\usage{
adjusted_accuracy(y, p)
}
\arguments{
\item{y}{true labels.}

\item{p}{predicted labels.}
}
\value{
adjusted accuracy (HR + CR / 2)
}
\description{
Compute adjusted accuracy given true and predicted values.
}
\examples{
y <- c(1,1,1,1,0,0,0,0,0,0,0,0,0)
p <- c(0,1,1,0,0,1,0,0,0,0,0,0,0)
mean(y == p) # accuracy
adjusted_accuracy(y, p)
y <- c(1,1,1,1,0,0,0,0,0,0,0,0,0)
p <- c(0,0,0,0,0,0,0,0,0,0,0,0,0)
mean(y == p) # accuracy
adjusted_accuracy(y, p)
}
